Archive the project
Thank you for using this. However, it is the right time to say goodbye to this project. We will NOT support this in future.

The new streaming API will be supported by JKook in version 0.53.0! Feel free to use that instead.
See the officially supported new streaming API here: https://developer.kookapp.cn/doc/http/voice